requirements.txt
tensorflow==2.9.2  # Adjust to the TensorFlow version installed
numpy==1.23.4       # Adjust to your numpy version
pillow==10.4.0       # Adjust to your pillow version
matplotlib==3.9.2   # Adjust to your matplotlib version
scikit-learn==1.5.2 # Adjust to your scikit-learn version
